IndexedDB: Replace static_cast<size_t>(123) with 123u in tests/checks
R=dgrogan@chromium.org

Review URL: https://codereview.chromium.org/143733005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@247006 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4532a305ff8a08593b02b973a25685368c35cb64

+++ content/browser/indexed_db/indexed_db_database.cc
@@ -1317,7 +1317,7 @@ void IndexedDBDatabase::ProcessPendingCalls() {
                                      pending_call->ReleaseConnection(),
                                      pending_call->transaction_id(),
                                      pending_call->version());
-    DCHECK_EQ(static_cast<size_t>(1), ConnectionCount());
+    DCHECK_EQ(1u, ConnectionCount());
     // Fall through would be a no-op, since transaction must complete
     // asynchronously.
     DCHECK(IsDeleteDatabaseBlocked());
